test: centralize duplicated test helpers into test-utils (#488)
